AI news Assistant/requirements.txt
streamlit
python-dotenv
requests
langchain
openai
apscheduler
